Create AbstractCreateHandle class and make HoodieCreateHandle extend it
